+++ examples/base/src/pages/Demo/Detail/index.tsx
@@ -23,6 +23,7 @@ const Detail = ({ updateData, onSearch }: DetailProps) => {
 
   const form = useForm()
   const form2 = useForm()
+
   const onClose = () => dispatch({ type: 'demo/clean' })
 
   // eslint-disable-next-line no-unused-vars
@@ -64,14 +65,15 @@ const Detail = ({ updateData, onSearch }: DetailProps) => {
   }
   const handleSave = async () => {
     // 触发验证
-    await form?.submitvalidate()
-    await form2?.submitvalidate()
+    await form.submitvalidate()
+    await form2.submitvalidate()
     // 获取错误信息
     const errors = form.getErrors()
     const errors2 = form2.getErrors()
 
     if (errors && Object.keys(errors).length > 0) return
     if (errors2 && Object.keys(errors2).length > 0) return
+
     mutate()
   }
 
@@ -148,6 +150,7 @@ const Detail = ({ updateData, onSearch }: DetailProps) => {
         buttonsContainer={{ justifyContent: 'flex-start' }}
         formDatas={items2(queryInfo) as any}
       />
+      <div style={{ marginTop: 10 }} />
     </ProDrawer>
   )
 }