+++ packages/components/src/ProForm/hooks/useForm.tsx
@@ -1,24 +1,35 @@
-import { useRef } from 'react';
-import { FormRefProps } from '../type';
+import { useState } from 'react';
+import { FormListProps } from '../type';
 
 const useForm = () => {
-  const formRef = useRef<FormRefProps>();
+  // 表单组件实例
+  const [formInstance, setFormInstance] = useState<any>(null);
 
-  // 触发验证表单
-  const submitvalidate = () => formRef.current?.submitvalidate() || null;
-  // 重置表单
-  const resetForm = () => formRef.current?.resetForm() || null;
+  // 表单报错信息
+  const [errors, setErrors] = useState<any>(null);
+
+  // 多表单列表
+  const [formList, setFormList] = useState<{ current?: FormListProps }[]>([]);
+
+  // 触发验证表单兼容原有api验证
+  const submitvalidate = () => formInstance?.current?.onSubmit();
+
   // 获取表单值
-  const getFieldValues = () => formRef.current?.getFieldValues() || null;
+  const getFieldValues = () => formInstance.current?.getFieldValues() || {};
+
   // 获取错误信息
-  const getErrors = () => formRef.current?.errors || {};
+  const getErrors = () => errors;
 
   return {
-    formRef,
     submitvalidate,
-    resetForm,
-    getFieldValues,
     getErrors,
+    getFieldValues,
+    // 多表单列表
+    formList,
+    // --------------
+    setFormInstance,
+    setErrors,
+    setFormList,
   };
 };